Make use of swagger plugin for listings query params and response model
