+++ backend/core/src/listings/listings.controller.ts
@@ -1,6 +1,7 @@
 import { Controller, Get, Query } from "@nestjs/common"
-import { ListingsFindAllResponse, ListingsQueryParams, ListingsService } from "./listings.service"
+import { ListingsService } from "./listings.service"
 import { ApiOkResponse } from "@nestjs/swagger"
+import { ListingsFindAllQueryParams, ListingsFindAllResponse } from "./listings.dto"
 
 @Controller()
 export class ListingsController {
@@ -11,7 +12,7 @@ export class ListingsController {
     type: ListingsFindAllResponse,
   })
   public async getAll(
-    @Query() queryParams?: ListingsQueryParams
+    @Query() queryParams?: ListingsFindAllQueryParams
   ): Promise<ListingsFindAllResponse> {
     return await this.listingsService.findAll(queryParams)
   }
