+++ backend/core/src/listings/listings.dto.ts
@@ -1,0 +1,12 @@
+import { ListingEntity } from "../entity/listing.entity"
+import { ListingsResponseStatus } from "./listings.service"
+
+export class ListingsFindAllQueryParams {
+  jsonpath?: string
+}
+
+export class ListingsFindAllResponse {
+  status: ListingsResponseStatus
+  listings: ListingEntity[]
+  amiCharts: any
+}
